Moved validate and test_validate to recipe module.
Corresponding changes to other files.
