+++ workflows/recipe/__init__.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 from workflows.recipe.recipe import Recipe
 from workflows.recipe.wrapper import RecipeWrapper
+from workflows.recipe.validate import validate_recipe
 
 
 def _wrap_subscription(
