+++ workflows/recipe/test_validate.py
@@ -5,7 +5,7 @@ Tests the functionality of validate.py with several different recipes which shou
 import pytest
 import mock
 import workflows
-from workflows.validate import validate_recipe, main
+from workflows.recipe.validate import validate_recipe, main
 import sys
 
 # Python 2 and 3 compatibility
@@ -98,7 +98,7 @@ def test_workflows_error_when_validating_incorrect_workflows_recipe():
 
 
 # Create a mock of the validate call
-@mock.patch("workflows.validate.validate_recipe")
+@mock.patch("workflows.recipe.validate.validate_recipe")
 def test_command_line_validation_one_argument(mock_requests):
     # Create fake arguments to test on
     test_args = ["validate.py", "file1"]
@@ -116,7 +116,7 @@ def test_exit_when_command_line_validation_given_no_arguments():
 
 
 # Create a mock of the validate call
-@mock.patch("workflows.validate.validate_recipe")
+@mock.patch("workflows.recipe.validate.validate_recipe")
 def test_command_line_validation_multiple_arguments(mock_requests):
     # Create fake arguments to test on
     test_args = ["validate.py", "file1", "file2", "file3"]
@@ -128,7 +128,7 @@ def test_command_line_validation_multiple_arguments(mock_requests):
     assert mock_requests.call_count == 3
 
 
-@mock.patch("workflows.validate.validate_recipe")
+@mock.patch("workflows.recipe.validate.validate_recipe")
 def test_system_exit_when_error_raised_from_command_line_validation(mock_requests):
     mock_requests.get.side_effect = Exception
     # Create fake arguments to test on
